Move composed scenario authority to ordinary producers
