+++ docs/reference/operation-invocation.md
@@ -12,6 +12,9 @@ Typed invocation for a routed AW action. This object owns machine action identit
 | --- | --- | --- | --- | --- | --- | --- |
 | (root) | object | yes |  | Typed invocation for a routed AW action. This object owns machine action identity; rendered commands are display or manual transport forms. |  | x-agentic-workspace-doc-role: "contract-reference" |
 | `kind` | const `"agentic-workspace/operation-invocation/v1"` | yes |  | Schema discriminator for typed operation invocations. |  |  |
+| `producer_module` | string | no |  | Canonical module that produced the typed invocation. |  |  |
+| `producer_function` | string | no |  | Canonical function that produced the typed invocation. |  |  |
+| `producer_revision` | string | no |  | Revision of the producer-owned invocation inputs. |  |  |
 | `operation_id` | string | yes |  | Stable operation identity from the operation contract registry. |  |  |
 | `contract_version` | const `"agentic-workspace/operation/v1"` | yes |  | Version of the operation contract this invocation targets. |  |  |
 | `arguments` | object | yes |  | Typed operation arguments independent of shell rendering. |  |  |
